pyproject.toml
[tool.poetry]
name = "toobee"
version = "0.1.0"
description = "Too bee or not too bee?"
authors = ["Dylan Ferris <dylan@psilly.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyzmq = "^23.2.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
scriv = "^0.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
